[DataGrid] Make rowCount part of grid state
