+++ docs/pages/x/api/data-grid/data-grid-pro.json
@@ -402,6 +402,10 @@
         "describedArgs": ["params", "event", "details"]
       }
     },
+    "onRowCountChange": {
+      "type": { "name": "func" },
+      "signature": { "type": "function(count: number) => void", "describedArgs": ["count"] }
+    },
     "onRowDoubleClick": {
       "type": { "name": "func" },
       "signature": {
