+++ docs/pages/x/api/data-grid/grid-api.md
@@ -116,6 +116,7 @@ import { GridApi } from '@mui/x-data-grid';
 | <span class="prop-name">setPinnedColumns [<span class="plan-pro" title="Pro plan"></span>](/x/introduction/licensing/#pro-plan)</span>                        | <span class="prop-type">(pinnedColumns: GridPinnedColumnFields) =&gt; void</span>                                                                                                                        | Changes the pinned columns.                                                                                                                                                                                                     |
 | <span class="prop-name">setQuickFilterValues</span>                                                                                                           | <span class="prop-type">(values: any[]) =&gt; void</span>                                                                                                                                                | Set the quick filter values to the one given by `values`                                                                                                                                                                        |
 | <span class="prop-name">setRowChildrenExpansion [<span class="plan-pro" title="Pro plan"></span>](/x/introduction/licensing/#pro-plan)</span>                 | <span class="prop-type">(id: GridRowId, isExpanded: boolean) =&gt; void</span>                                                                                                                           | Expand or collapse a row children.                                                                                                                                                                                              |
+| <span class="prop-name">setRowCount</span>                                                                                                                    | <span class="prop-type">(rowCount: number) =&gt; void</span>                                                                                                                                             | Sets the `rowCount` to a new value.                                                                                                                                                                                             |
 | <span class="prop-name">setRowGroupingCriteriaIndex [<span class="plan-premium" title="Premium plan"></span>](/x/introduction/licensing/#premium-plan)</span> | <span class="prop-type">(groupingCriteriaField: string, groupingIndex: number) =&gt; void</span>                                                                                                         | Sets the grouping index of a grouping criteria.                                                                                                                                                                                 |
 | <span class="prop-name">setRowGroupingModel [<span class="plan-premium" title="Premium plan"></span>](/x/introduction/licensing/#premium-plan)</span>         | <span class="prop-type">(model: GridRowGroupingModel) =&gt; void</span>                                                                                                                                  | Sets the columns to use as grouping criteria.                                                                                                                                                                                   |
 | <span class="prop-name">setRowIndex [<span class="plan-pro" title="Pro plan"></span>](/x/introduction/licensing/#pro-plan)</span>                             | <span class="prop-type">(rowId: GridRowId, targetIndex: number) =&gt; void</span>                                                                                                                        | Moves a row from its original position to the position given by `targetIndex`.                                                                                                                                                  |
