+++ docs/pages/x/api/data-grid/grid-pagination-model-api.json
@@ -1,0 +1,21 @@
+{
+  "name": "GridPaginationModelApi",
+  "description": "The pagination model API interface that is available in the grid `apiRef`.",
+  "properties": [
+    {
+      "name": "setPage",
+      "description": "Sets the displayed page to the value given by <code>page</code>.",
+      "type": "(page: number) => void"
+    },
+    {
+      "name": "setPageSize",
+      "description": "Sets the number of displayed rows to the value given by <code>pageSize</code>.",
+      "type": "(pageSize: number) => void"
+    },
+    {
+      "name": "setPaginationModel",
+      "description": "Sets the <code>paginationModel</code> to a new value.",
+      "type": "(model: GridPaginationModel) => void"
+    }
+  ]
+}
