+++ docs/pages/x/api/data-grid/grid-pagination-row-count-api.json
@@ -1,0 +1,11 @@
+{
+  "name": "GridPaginationRowCountApi",
+  "description": "The pagination row count API interface that is available in the grid `apiRef`.",
+  "properties": [
+    {
+      "name": "setRowCount",
+      "description": "Sets the <code>rowCount</code> to a new value.",
+      "type": "(rowCount: number) => void"
+    }
+  ]
+}
