+++ docs/scripts/api/buildInterfacesDocumentation.ts
@@ -48,7 +48,8 @@ const GRID_API_INTERFACES_WITH_DEDICATED_PAGES = [
   'GridEditingApi',
   'GridExcelExportApi',
   'GridFilterApi',
-  'GridPaginationApi',
+  'GridPaginationModelApi',
+  'GridPaginationRowCountApi',
   'GridPrintExportApi',
   'GridRowGroupingApi',
   'GridRowMultiSelectionApi',
