+++ docs/translations/api-docs/data-grid/data-grid/data-grid.json
@@ -337,6 +337,10 @@
         "details": "Additional details for this callback."
       }
     },
+    "onRowCountChange": {
+      "description": "Callback fired when the row count has changed.",
+      "typeDescriptions": { "count": "Updated row count." }
+    },
     "onRowDoubleClick": {
       "description": "Callback fired when a double click event comes from a row container element.",
       "typeDescriptions": {
