+++ packages/x-data-grid/src/hooks/features/pagination/gridPaginationInterfaces.ts
@@ -2,16 +2,18 @@ import { GridPaginationModel } from '../../../models/gridPaginationProps';
 
 export interface GridPaginationState {
   paginationModel: GridPaginationModel;
+  rowCount: number;
 }
 
 export interface GridPaginationInitialState {
   paginationModel?: Partial<GridPaginationModel>;
+  rowCount?: number;
 }
 
 /**
- * The pagination API interface that is available in the grid `apiRef`.
+ * The pagination model API interface that is available in the grid `apiRef`.
  */
-export interface GridPaginationApi {
+export interface GridPaginationModelApi {
   /**
    * Sets the displayed page to the value given by `page`.
    * @param {number} page The new page number.
@@ -27,4 +29,15 @@ export interface GridPaginationApi {
    * @param {GridPaginationModel} model The new model value.
    */
   setPaginationModel: (model: GridPaginationModel) => void;
+}
+
+/**
+ * The pagination row count API interface that is available in the grid `apiRef`.
+ */
+export interface GridPaginationRowCountApi {
+  /**
+   * Sets the `rowCount` to a new value.
+   * @param {number} rowCount The new row count value.
+   */
+  setRowCount: (rowCount: number) => void;
 }
