+++ packages/x-data-grid/src/hooks/features/pagination/index.ts
@@ -1,6 +1,7 @@
 export * from './gridPaginationSelector';
 export type {
-  GridPaginationApi,
+  GridPaginationModelApi,
+  GridPaginationRowCountApi,
   GridPaginationState,
   GridPaginationInitialState,
 } from './gridPaginationInterfaces';
