+++ packages/x-data-grid/src/models/props/DataGridProps.ts
@@ -592,6 +592,11 @@ export interface DataGridPropsWithoutDefaultValue<R extends GridValidRowModel = 
    */
   onPaginationModelChange?: (model: GridPaginationModel, details: GridCallbackDetails) => void;
   /**
+   * Callback fired when the row count has changed.
+   * @param {number} count Updated row count.
+   */
+  onRowCountChange?: (count: number) => void;
+  /**
    * Callback fired when the preferences panel is closed.
    * @param {GridPreferencePanelParams} params With all properties from [[GridPreferencePanelParams]].
    * @param {MuiEvent<{}>} event The event object.
