+++ scripts/x-data-grid-premium.exports.json
@@ -404,10 +404,12 @@
   { "name": "gridPaginatedVisibleSortedGridRowEntriesSelector", "kind": "Variable" },
   { "name": "gridPaginatedVisibleSortedGridRowIdsSelector", "kind": "Variable" },
   { "name": "GridPagination", "kind": "Variable" },
-  { "name": "GridPaginationApi", "kind": "Interface" },
   { "name": "GridPaginationInitialState", "kind": "Interface" },
   { "name": "GridPaginationModel", "kind": "Interface" },
+  { "name": "GridPaginationModelApi", "kind": "Interface" },
   { "name": "gridPaginationModelSelector", "kind": "Variable" },
+  { "name": "GridPaginationRowCountApi", "kind": "Interface" },
+  { "name": "gridPaginationRowCountSelector", "kind": "Variable" },
   { "name": "gridPaginationRowRangeSelector", "kind": "Variable" },
   { "name": "gridPaginationSelector", "kind": "Variable" },
   { "name": "GridPaginationState", "kind": "Interface" },
